# sql_app/openapi.json
{
   "openapi":"3.0.2",
   "info":{
      "title":"FastAPI",
      "version":"0.1.0"
   },
   "paths":{
      "/employee/":{
         "post":{
            "summary":"Create Employee",
            "operationId":"create_employee_employee__post",
            "requestBody":{
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/EmployeeBase"
                     }
                  }
               },
               "required":true
            },
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/Employee"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/employees/":{
         "get":{
            "summary":"Read Employees",
            "operationId":"read_employees_employees__get",
            "parameters":[
               {
                  "required":false,
                  "schema":{
                     "title":"Skip",
                     "type":"integer",
                     "default":0
                  },
                  "name":"skip",
                  "in":"query"
               },
               {
                  "required":false,
                  "schema":{
                     "title":"Limit",
                     "type":"integer",
                     "default":100
                  },
                  "name":"limit",
                  "in":"query"
               }
            ],
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "title":"Response Read Employees Employees  Get",
                           "type":"array",
                           "items":{
                              "$ref":"#/components/schemas/Employee"
                           }
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/employees/{emp_id}":{
         "get":{
            "summary":"Read Employee",
            "operationId":"read_employee_employees__emp_id__get",
            "parameters":[
               {
                  "required":true,
                  "schema":{
                     "title":"Emp Id",
                     "type":"integer"
                  },
                  "name":"emp_id",
                  "in":"path"
               }
            ],
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/Employee"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/employees/{emp_id}/timesheet/":{
         "get":{
            "summary":"Read Timesheet For Emp",
            "operationId":"read_timesheet_for_emp_employees__emp_id__timesheet__get",
            "parameters":[
               {
                  "required":true,
                  "schema":{
                     "title":"Emp Id",
                     "type":"integer"
                  },
                  "name":"emp_id",
                  "in":"path"
               },
               {
                  "required":false,
                  "schema":{
                     "title":"Skip",
                     "type":"integer",
                     "default":0
                  },
                  "name":"skip",
                  "in":"query"
               },
               {
                  "required":false,
                  "schema":{
                     "title":"Limit",
                     "type":"integer",
                     "default":100
                  },
                  "name":"limit",
                  "in":"query"
               }
            ],
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "title":"Response Read Timesheet For Emp Employees  Emp Id  Timesheet  Get",
                           "type":"array",
                           "items":{
                              "$ref":"#/components/schemas/Timesheet"
                           }
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "post":{
            "summary":"Create Timesheet For Emp",
            "operationId":"create_timesheet_for_emp_employees__emp_id__timesheet__post",
            "parameters":[
               {
                  "required":true,
                  "schema":{
                     "title":"Emp Id",
                     "type":"integer"
                  },
                  "name":"emp_id",
                  "in":"path"
               }
            ],
            "requestBody":{
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/TimesheetCreate"
                     }
                  }
               },
               "required":true
            },
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/Timesheet"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/employees/{emp_id}/timesheet/{timesheet_date}":{
         "get":{
            "summary":"Read Timesheet For Emp",
            "operationId":"read_timesheet_for_emp_employees__emp_id__timesheet__timesheet_date__get",
            "parameters":[
               {
                  "required":true,
                  "schema":{
                     "title":"Emp Id",
                     "type":"integer"
                  },
                  "name":"emp_id",
                  "in":"path"
               },
               {
                  "required":true,
                  "schema":{
                     "title":"Timesheet Date",
                     "type":"string"
                  },
                  "name":"timesheet_date",
                  "in":"path"
               },
               {
                  "required":false,
                  "schema":{
                     "title":"Skip",
                     "type":"integer",
                     "default":0
                  },
                  "name":"skip",
                  "in":"query"
               },
               {
                  "required":false,
                  "schema":{
                     "title":"Limit",
                     "type":"integer",
                     "default":100
                  },
                  "name":"limit",
                  "in":"query"
               }
            ],
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "title":"Response Read Timesheet For Emp Employees  Emp Id  Timesheet  Timesheet Date  Get",
                           "type":"array",
                           "items":{
                              "$ref":"#/components/schemas/Timesheet"
                           }
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      }
   },
   "components":{
      "schemas":{
         "Employee":{
            "title":"Employee",
            "required":[
               "name",
               "id"
            ],
            "type":"object",
            "properties":{
               "name":{
                  "title":"Name",
                  "type":"string"
               },
               "id":{
                  "title":"Id",
                  "type":"integer"
               }
            }
         },
         "EmployeeBase":{
            "title":"EmployeeBase",
            "required":[
               "name"
            ],
            "type":"object",
            "properties":{
               "name":{
                  "title":"Name",
                  "type":"string"
               }
            }
         },
         "HTTPValidationError":{
            "title":"HTTPValidationError",
            "type":"object",
            "properties":{
               "detail":{
                  "title":"Detail",
                  "type":"array",
                  "items":{
                     "$ref":"#/components/schemas/ValidationError"
                  }
               }
            }
         },
         "Timesheet":{
            "title":"Timesheet",
            "required":[
               "date",
               "amount_of_hours",
               "id",
               "emp_id"
            ],
            "type":"object",
            "properties":{
               "date":{
                  "title":"Date",
                  "type":"string"
               },
               "amount_of_hours":{
                  "title":"Amount Of Hours",
                  "type":"integer"
               },
               "id":{
                  "title":"Id",
                  "type":"integer"
               },
               "emp_id":{
                  "title":"Emp Id",
                  "type":"integer"
               }
            }
         },
         "TimesheetCreate":{
            "title":"TimesheetCreate",
            "required":[
               "date",
               "amount_of_hours"
            ],
            "type":"object",
            "properties":{
               "date":{
                  "title":"Date",
                  "type":"string"
               },
               "amount_of_hours":{
                  "title":"Amount Of Hours",
                  "type":"integer"
               }
            }
         },
         "ValidationError":{
            "title":"ValidationError",
            "required":[
               "loc",
               "msg",
               "type"
            ],
            "type":"object",
            "properties":{
               "loc":{
                  "title":"Location",
                  "type":"array",
                  "items":{
                     "anyOf":[
                        {
                           "type":"string"
                        },
                        {
                           "type":"integer"
                        }
                     ]
                  }
               },
               "msg":{
                  "title":"Message",
                  "type":"string"
               },
               "type":{
                  "title":"Error Type",
                  "type":"string"
               }
            }
         }
      }
   }
}